+++ tests/modules/test_SeleniumExecutor.py
@@ -1,14 +1,15 @@
+import csv
 import os
 import shutil
 import time
-import csv
+
 import yaml
 
-from bzt.engine import ScenarioExecutor
-from tests import setup_test_logging, BZTestCase, local_paths_config, __dir__
+from bzt.engine import ScenarioExecutor, Provisioning
 from bzt.modules.selenium import SeleniumExecutor, JUnitJar
-from tests.mocks import EngineEmul
 from bzt.utils import BetterDict
+from tests import setup_test_logging, BZTestCase, local_paths_config, __dir__
+from tests.mocks import EngineEmul
 
 setup_test_logging()
 
@@ -58,11 +59,14 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
 
         obj = self.get_selenium_executor()
         obj.settings.merge({"selenium-tools": {
-            "junit": {"selenium-server": os.path.join(dummy_installation_path, "selenium-server.jar")}}})
+            "junit": {"selenium-server": os.path.join(dummy_installation_path, "selenium-server.jar")}
+        }})
         obj.settings.merge({"selenium-tools": {
-            "junit": {"hamcrest-core": os.path.join(dummy_installation_path, "tools", "junit", "hamcrest-core.jar")}}})
+            "junit": {"hamcrest-core": os.path.join(dummy_installation_path, "tools", "junit", "hamcrest-core.jar")}
+        }})
         obj.settings.merge({"selenium-tools": {
-            "junit": {"path": os.path.join(dummy_installation_path, "tools", "junit", "junit.jar")}}})
+            "junit": {"path": os.path.join(dummy_installation_path, "tools", "junit", "junit.jar")}
+        }})
 
         obj.execution = BetterDict()
         obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/jar/"}})
@@ -87,8 +91,7 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
         :return:
         """
         obj = self.get_selenium_executor()
-        obj.execution.merge(
-            {"scenario": {"script": __dir__() + "/../selenium/java/TestBlazemeterFail.java"}})
+        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/java/TestBlazemeterFail.java"}})
         obj.prepare()
         self.assertTrue(os.path.exists(os.path.join(obj.runner.working_dir, "TestBlazemeterFail.java")))
         self.assertTrue(os.path.exists(os.path.join(obj.runner.working_dir, "TestBlazemeterFail.class")))
@@ -118,8 +121,7 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
         """
         obj = self.get_selenium_executor()
         obj.execution = BetterDict()
-        obj.execution.merge(
-            {"scenario": {"script": __dir__() + "/../selenium/java_package/"}})
+        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/java_package/"}})
         obj.prepare()
         self.assertTrue(os.path.exists(os.path.join(obj.runner.working_dir, "compiled.jar")))
 
@@ -129,9 +131,13 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
         :return:
         """
         obj = self.get_selenium_executor()
-        obj.engine.config.merge(
-            {'execution': {'scenario': {'script': __dir__() + '/../selenium/java_package/'}, 'executor': 'selenium'},
-             'reporting': [{'module': 'junit-xml'}]})
+        obj.engine.config.merge({
+            'execution': {
+                'scenario': {'script': __dir__() + '/../selenium/java_package/'},
+                'executor': 'selenium'
+            },
+            'reporting': [{'module': 'junit-xml'}]
+        })
         obj.engine.config.merge({"provisioning": "local"})
         obj.execution = obj.engine.config['execution']
         obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
@@ -149,11 +155,9 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
         """
         obj = self.get_selenium_executor()
         obj.execution = BetterDict()
-        obj.execution.merge(
-            {"scenario": {"script": __dir__() + "/../selenium/jar/dummy.jar"}})
+        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/jar/dummy.jar"}})
         obj.prepare()
-        self.assertTrue(
-            os.path.exists(os.path.join(obj.runner.working_dir, "dummy.jar")))
+        self.assertTrue(os.path.exists(os.path.join(obj.runner.working_dir, "dummy.jar")))
 
     def test_prepare_jar_folder(self):
         """
@@ -173,13 +177,16 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
         :return:
         """
         obj = self.get_selenium_executor()
-        obj.engine.config.merge(
-            {'execution': {'scenario': {'script': __dir__() + '/../selenium/jar/'}, 'executor': 'selenium'},
-             'reporting': [{'module': 'junit-xml'}]})
+        obj.engine.config.merge({
+            'execution': {
+                'scenario': {'script': __dir__() + '/../selenium/jar/'},
+                'executor': 'selenium'
+            },
+            'reporting': [{'module': 'junit-xml'}]
+        })
         obj.engine.config.merge({"provisioning": "local"})
         obj.execution = obj.engine.config['execution']
-        obj.execution.merge(
-            {"scenario": {"script": __dir__() + "/../selenium/jar/dummy.jar"}})
+        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/jar/dummy.jar"}})
         obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
         obj.prepare()
         obj.startup()
@@ -202,9 +209,13 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
         :return:
         """
         obj = self.get_selenium_executor()
-        obj.engine.config.merge(
-            {'execution': {'scenario': {'script': __dir__() + '/../selenium/jar/'}, 'executor': 'selenium'},
-             'reporting': [{'module': 'junit-xml'}]})
+        obj.engine.config.merge({
+            'execution': {
+                'scenario': {'script': __dir__() + '/../selenium/jar/'},
+                'executor': 'selenium'
+            },
+            'reporting': [{'module': 'junit-xml'}]
+        })
         obj.engine.config.merge({"provisioning": "local"})
         obj.execution = obj.engine.config['execution']
         obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
@@ -229,13 +240,16 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
         :return:
         """
         obj = self.get_selenium_executor()
-        obj.engine.config.merge(
-            {'execution': {'scenario': {'script': __dir__() + '/../selenium/java/'}, 'executor': 'selenium'},
-             'reporting': [{'module': 'junit-xml'}]})
+        obj.engine.config.merge({
+            'execution': {
+                'scenario': {'script': __dir__() + '/../selenium/java/'},
+                'executor': 'selenium'
+            },
+            'reporting': [{'module': 'junit-xml'}]
+        })
         obj.engine.config.merge({"provisioning": "local"})
         obj.execution = obj.engine.config['execution']
-        obj.execution.merge(
-            {"scenario": {"script": __dir__() + "/../selenium/java/TestBlazemeterFail.java"}})
+        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/java/TestBlazemeterFail.java"}})
         obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
         obj.prepare()
         obj.startup()
@@ -259,9 +273,13 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
         :return:
         """
         obj = self.get_selenium_executor()
-        obj.engine.config.merge(
-            {'execution': {'scenario': {'script': __dir__() + '/../selenium/java/'}, 'executor': 'selenium'},
-             'reporting': [{'module': 'junit-xml'}]})
+        obj.engine.config.merge({
+            'execution': {
+                'scenario': {'script': __dir__() + '/../selenium/java/'},
+                'executor': 'selenium'
+            },
+            'reporting': [{'module': 'junit-xml'}]
+        })
         obj.engine.config.merge({"provisioning": "local"})
         obj.execution = obj.engine.config['execution']
         obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
@@ -288,9 +306,13 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
         """
         obj = self.get_selenium_executor()
         obj.engine.config = BetterDict()
-        obj.engine.config.merge(
-            {ScenarioExecutor.EXEC: {"executor": "selenium",
-                                     "scenario": {"script": __dir__() + "/../selenium/invalid/NotJUnittest.java"}}})
+        obj.engine.config.merge({
+            Provisioning.PROV: "local",
+            ScenarioExecutor.EXEC: {
+                "executor": "selenium",
+                "scenario": {"script": __dir__() + "/../selenium/invalid/NotJUnittest.java"}
+            }
+        })
         obj.execution = obj.engine.config['execution']
         obj.prepare()
         obj.startup()
@@ -304,9 +326,13 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
 
     def test_resource_files_collection_remote_java(self):
         obj = self.get_selenium_executor()
-        obj.engine.config.merge(
-            {'execution': {'scenario': {'script': __dir__() + '/../selenium/java/'}, 'executor': 'selenium'},
-             'reporting': [{'module': 'junit-xml'}]})
+        obj.engine.config.merge({
+            'execution': {
+                'scenario': {'script': __dir__() + '/../selenium/java/'},
+                'executor': 'selenium'
+            },
+            'reporting': [{'module': 'junit-xml'}]
+        })
         obj.engine.config.merge({"provisioning": "local"})
         obj.execution = obj.engine.config['execution']
         obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
@@ -317,9 +343,13 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
 
     def test_resource_files_collection_remote_jar(self):
         obj = self.get_selenium_executor()
-        obj.engine.config.merge(
-            {'execution': {'scenario': {'script': __dir__() + '/../selenium/jar/'}, 'executor': 'selenium'},
-             'reporting': [{'module': 'junit-xml'}]})
+        obj.engine.config.merge({
+            'execution': {
+                'scenario': {'script': __dir__() + '/../selenium/jar/'},
+                'executor': 'selenium'
+            },
+            'reporting': [{'module': 'junit-xml'}]
+        })
         obj.engine.config.merge({"provisioning": "local"})
         obj.execution = obj.engine.config['execution']
         obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
@@ -339,7 +369,8 @@ class TestSeleniumNoseRunner(BZTestCase):
         obj.engine = EngineEmul()
         obj.execution = BetterDict()
         obj.execution.merge({"scenario": {
-            "script": __dir__() + "/../selenium/python/test_blazemeter_fail.py"}})
+            "script": __dir__() + "/../selenium/python/test_blazemeter_fail.py"
+        }})
         obj.prepare()
         python_scripts = os.listdir(obj.runner.working_dir)
         self.assertEqual(len(python_scripts), 1)
@@ -366,14 +397,19 @@ class TestSeleniumNoseRunner(BZTestCase):
         obj = SeleniumExecutor()
         obj.engine = EngineEmul()
         obj.engine.config = BetterDict()
-        obj.engine.config.merge(
-            {'execution': {'scenario': {'script': __dir__() + '/../selenium/python/'}, 'executor': 'selenium'},
-             'reporting': [{'module': 'junit-xml'}]})
+        obj.engine.config.merge({
+            'execution': {
+                'scenario': {'script': __dir__() + '/../selenium/python/'},
+                'executor': 'selenium'
+            },
+            'reporting': [{'module': 'junit-xml'}]
+        })
         obj.engine.config.merge({"provisioning": "local"})
         obj.execution = obj.engine.config['execution']
 
         obj.execution.merge({"scenario": {
-            "script": __dir__() + "/../selenium/python/test_blazemeter_fail.py"}})
+            "script": __dir__() + "/../selenium/python/test_blazemeter_fail.py"
+        }})
 
         obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
         obj.prepare()
@@ -394,9 +430,13 @@ class TestSeleniumNoseRunner(BZTestCase):
         obj = SeleniumExecutor()
         obj.engine = EngineEmul()
         obj.engine.config = BetterDict()
-        obj.engine.config.merge(
-            {'execution': {'scenario': {'script': __dir__() + '/../selenium/python/'}, 'executor': 'selenium'},
-             'reporting': [{'module': 'junit-xml'}]})
+        obj.engine.config.merge({
+            'execution': {
+                'scenario': {'script': __dir__() + '/../selenium/python/'},
+                'executor': 'selenium'
+            },
+            'reporting': [{'module': 'junit-xml'}]
+        })
         obj.engine.config.merge({"provisioning": "local"})
         obj.execution = obj.engine.config['execution']
         obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
@@ -418,9 +458,12 @@ class TestSeleniumNoseRunner(BZTestCase):
         obj = SeleniumExecutor()
         obj.engine = EngineEmul()
         obj.engine.config = BetterDict()
-        obj.engine.config.merge(
-            {ScenarioExecutor.EXEC: {"executor": "selenium",
-                                     "scenario": {"script": __dir__() + "/../selenium/invalid/dummy.py"}}})
+        obj.engine.config.merge({
+            ScenarioExecutor.EXEC: {
+                "executor": "selenium",
+                "scenario": {"script": __dir__() + "/../selenium/invalid/dummy.py"}
+            }
+        })
         obj.execution = obj.engine.config['execution']
         obj.prepare()
         obj.startup()
@@ -466,9 +509,12 @@ class TestSeleniumStuff(SeleniumTestCase):
         obj.engine = self.engine_obj
         obj.settings = self.selenium_config
         obj.engine.config = BetterDict()
-        obj.engine.config.merge(
-            {ScenarioExecutor.EXEC: {"executor": "selenium",
-                                     "scenario": {"script": __dir__() + "/../selenium/invalid/invalid.java"}}})
+        obj.engine.config.merge({
+            ScenarioExecutor.EXEC: {
+                "executor": "selenium",
+                "scenario": {"script": __dir__() + "/../selenium/invalid/invalid.java"}
+            }
+        })
         obj.execution = obj.engine.config['execution']
         self.assertRaises(RuntimeError, obj.prepare)
 
@@ -480,9 +526,10 @@ class TestSeleniumStuff(SeleniumTestCase):
         obj = SeleniumExecutor()
         obj.engine = EngineEmul()
         obj.engine.config = BetterDict()
-        obj.engine.config.merge(
-            {ScenarioExecutor.EXEC: {"executor": "selenium",
-                                     "scenario": {"script": __dir__() + "/../selenium/invalid/not_found"}}})
+        obj.engine.config.merge({ScenarioExecutor.EXEC: {
+            "executor": "selenium",
+            "scenario": {"script": __dir__() + "/../selenium/invalid/not_found"}
+        }})
         obj.execution = obj.engine.config['execution']
         self.assertRaises(RuntimeError, obj.prepare)
 
@@ -494,9 +541,10 @@ class TestSeleniumStuff(SeleniumTestCase):
         obj = SeleniumExecutor()
         obj.engine = self.engine_obj
         obj.settings = self.selenium_config
-        obj.engine.config.merge(
-            {ScenarioExecutor.EXEC: {"executor": "selenium",
-                                     "scenario": {"script": __dir__() + "/../selenium/invalid/SeleniumTest.java"}}})
+        obj.engine.config.merge({ScenarioExecutor.EXEC: {
+            "executor": "selenium",
+            "scenario": {"script": __dir__() + "/../selenium/invalid/SeleniumTest.java"}
+        }})
         obj.execution = obj.engine.config['execution']
         obj.prepare()
         obj.startup()
@@ -516,9 +564,10 @@ class TestSeleniumStuff(SeleniumTestCase):
         obj = SeleniumExecutor()
         obj.engine = self.engine_obj
         obj.settings = self.selenium_config
-        obj.engine.config.merge(
-            {ScenarioExecutor.EXEC: {"executor": "selenium",
-                                     "scenario": {"script": __dir__() + "/../selenium/invalid/SimpleTest.java"}}})
+        obj.engine.config.merge({ScenarioExecutor.EXEC: {
+            "executor": "selenium",
+            "scenario": {"script": __dir__() + "/../selenium/invalid/SimpleTest.java"}
+        }})
         obj.execution = obj.engine.config['execution']
         obj.prepare()
         obj.startup()
@@ -539,8 +588,8 @@ class TestSeleniumStuff(SeleniumTestCase):
         obj.engine = self.engine_obj
         obj.settings = self.selenium_config
         obj.engine.config.merge({ScenarioExecutor.EXEC: {
-            "executor": "selenium", "scenario": {"script": __dir__() + "/../selenium/invalid/selenium1.java"}}
-        })
+            "executor": "selenium", "scenario": {"script": __dir__() + "/../selenium/invalid/selenium1.java"}
+        }})
         obj.execution = obj.engine.config['execution']
         obj.prepare()
         obj.startup()